server/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django_csp==3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
Markdown==3.5.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.1.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
wheel==0.42.0
